math: Move module to `lofty::util::math`
