+++ src/mp4/properties.rs
@@ -3,8 +3,8 @@ use super::read::{nested_atom, skip_unneeded, AtomReader};
 use crate::config::ParsingMode;
 use crate::error::{LoftyError, Result};
 use crate::macros::{decode_err, err, try_vec};
-use crate::math::RoundedDivision;
 use crate::properties::FileProperties;
+use crate::util::math::RoundedDivision;
 
 use std::io::{Cursor, Read, Seek, SeekFrom};
 use std::time::Duration;
