+++ src/ogg/opus/properties.rs
@@ -1,8 +1,8 @@
 use super::find_last_page;
 use crate::error::Result;
 use crate::macros::decode_err;
-use crate::math::RoundedDivision;
 use crate::properties::{ChannelMask, FileProperties};
+use crate::util::math::RoundedDivision;
 
 use std::io::{Read, Seek, SeekFrom};
 use std::time::Duration;
