+++ src/util/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod alloc;
+pub(crate) mod math;
 pub(crate) mod text;
